Tidy up - new directories
Instead of everything being stored in the single directory, the new directories were created: source/ (for the .cpp files), bin/ (for the executables) & output_files/ (for the output files created by the executables).  The files makefile and jobscript.sh are still in the main directory but have been updated to handle the new directories.  Also, the new modularised files (MomentCalculations.cpp, NegativityChecks.cpp, EntropyCalculations.cpp, MarginalCreation.cpp & EquilibriumSolution.cpp) were added to the repository, as they were not in the last commit.