Update to makefile
The makefile will now compile each .cpp file individually and then link them to form the executable.  This means that now, after editing a file, only the objects which depend on that file must be recompiled as opposed to recompiling everything every time.  This also means that header files have now been written for every .cpp file.  This code does indeed compile but there seems to be an issue with openMP meaning that it does NOT run as expected.