Individual compile then link working - running slower
The previous issue with openMP behaving incorrectly has now been fixed (the problem was with how IntM was declared as threadprivate).  This means that this version now compiles each .cpp file sperately and then links to create an executable which produces the same results as when everything compiled together.  It does, however, take twice the time to run and there also seems to be a subtle difference in the values at the level of machine accuracy (in particular the initial values of momentum, which were O(10^(-16)) are a little different, at the same order).  This may be related to the executable running sections of the code in a different order (or looping things unnecessarily that weren't before) and shall be investigated.