Added FieldCalculations files
Any functions previously in the advection_1.cpp file which involved calculations for the field were moved to a new file FieldCalculations.cpp and its appropriate header file was created.  This was mostly done in anticipation of changing the doping profile, but is also a bit of a tidy up.  Any files which contained references to those advection_1 functions now in FieldCalculations had their references fixed as well.