Added calculations for new doping profile
The FieldCalculations.cpp file was replaced with FieldCalculations_withND.cpp to change the constant doping profile to one in the form of a step function, with a higher constant value on the outsides and lower in the middle.  The code for the electric energy still need to be changed, but all calculations of the field and integrals of it have been replaced by the appropriate exact values for the new doping profile.