Changed conservation to only conserve mass
For the purpose of a non-uniform doping profile, the conservation routines have been edited so that the Lagrange minimisation problem only has the mass conservation constraint and so the resulting matrix involved is only 1xN^3.  Various variables had to have their dimensions changed in the process and a new file, called conservationRoutines_OnlyMass.cpp has replaced conservationRoutines.cpp.