Added code to compute linear Landau operator Q(f,M) for initial Maxwellian M
The functions ComputeQLinear, RK4Linear and ComputeDFTToMaxwellian were added to the collision_Routines1 file for running computations with the linear Landau operator Q(f,M), for the initial Maxwellian M and the code edited in the main file accordingly to allow the set up for this (by computing the FFT values of the initial condition and storing them in DFTMaxwell).  The functions generate_conv_weights2 and gHat3_2 were also added to the collision_Routines1 file to allow future expansion to multispecies code, where the species have different masses.  The declaration of the dielectric constant eps was also moved to the LP_ompi file instead of FieldCalculations as well.