Added LinearLandau macro
The option to use the macro LinearLandau has been added to LP_ompi.h to allow the code to easily switch between running the standard Landau operator and the linear version Q(f,M) for some initial Maxwellian.  Some extra info has been added to print to the ouput file, including a statement when the linear Landau is running and also the value of the dielectric constant epsilon being used.