Added MassConsOnly macro
The option to use the macro MassConsOnly has been added to LP_ompi.h to allow the code to easily switch between conserving all moments or just conserving the mass.  Note that this will now compile two files, both conservationRoutines.cpp (for full conservation) and conservationRoutines_OnlyMass.cpp (for only mass conservation), which both share the header file conservationRoutines.h.  The macro MassConsOnly means that actually only one of these cpp files has its content compiled though, depending on whether it is defined or not.  Various other changes had to be made to allow the option of switching back and forth.  It seems to be compiling but in order to continue testing, a different branch must be switched to.