Edited code to allow nu to have a different value on each cell.
The variable nu has been changed from a single double variable to a vector which allows a different value on each cell.  For the sake of labelling files and to test if the problem is collisionless, a new variable nu_max has been added  which stores the maximum value of nu on each cell (this is the value printed on the files and also if nu_max = 0 then the problem is collisionless, as nu > 0).  So far the program has only been tested when all values of nu are set to 0.05, but it gives the same results as when only a single value was used.