Added charge neutrality boundary conditions
The advection files now have a subroutine to implement charge neutrality boundary conditions.  By examining the marginals of the pdf, these seem to be the most appropriate boundary conditions for the simulations with an applied bias.  The calculations for the potential and field calculations also had some extra OpenMP code added and a slightly better implementation of setting the value of nthread was included.