Code running simulations for ions correctly
The issue with the simulation for ions was found and fixed (the flux involving the field had to be switched as ions travel the opposite way to electrons in a potential difference) so the code is now successfully running simulations for ions.  Macros were added so that the code can switch between modeling electrons or ions easily.