Slight tidy up
A few lines regarding the equilibrium solution have been commented out as they seem to be unnecessary - there is no need to calculate the relative entropy, only the entropy.  The code is also set up for simulating a wider doping profile, but the BCs used were periodic - fix this to the charge neutrality ones!