Added an option to run in the space homogeneous setting
Instead of needing to use a different set of code, the model for the simpler space homogeneous case has been added to the this version, along with a Boolean option to switch between the homogeneous and inhomogeneous settings.  It is definitely running but need to verify that the results agree with the original space homogeneous code.