Added a space homogeneous test
Included a fifth test to check the moments resulting from a space homogeneous run for five time-steps with gamma=-3.  Note that the moment file now has to print a 0 for the electric energy values so that the moment differ can operate in the same way as for the space inhomogeneous case.