Fixed some MPI printing issues
Forgot to restrict output of the read of gamma to just the rank 0 processor.  Also, the format of the space homogeneous moments file still was not quite right (it only printed 8 values, not the 9 which the inhomogeneous moments file does) - need to fix this for the space homogeneous test too.