Added option for GRVY input with different name
The code can now read in a GRVY input file with a different name by passing a command line argument after the name of the executable.  The output file now also prints the name of the executable that is being run and the name of the GRVY file being read from.