Fixed an issue that was causing a seg fault in the inhomogeneous version
The new marginal files that were added for the multispecies solutions didn't have enough memory allocated for the code to run with Homogeneous=False and there is still no space inhomogenous version of the multispecies code anyway so there was nothing to print.  This has now been fixed with a wrapper function for printing multispecies marginals.