Restored GRVY input option for NoField and set up make for small IC temp
For some reason the option for NoField was no longer available in the GRVY input file, but it has been put back.  Also changed the executable name because the variance of the Maxwellian in the Damping IC has been made small for hardsphere simulations.  This should be made into another GRVY input parsing variable though...