Added a DisparateMass option in the GRVY input file.
The code can now run in the single species case without any of the disparate mass variables being allocated by setting DisparateMass to False.  Unfortunately, when it is True, the single species variables are still being allocated for testing purposes.  About to add a DisparateMass_Debug variable to make this optional.