The DisparateMass variable now runs each case disjointly
If DisparateMass is False, only the single species variables are allocated.  If DisparateMass is True then only the disparate mass variables are allocated when DisparateMass_Check is False but, if DisparateMass_Check is True, all variables are allocated for comparison.  DisparateMass_Check is set to True by setting DM_Check to True under the DisparateMass header.