Fixed issue with momentum component blowing up
There was one dv that had not been changed in IntModes_L/_H that has been fixed.  The time-stepping still dramatically alters the mass and energy when the masses are different.  When the mass ratio is one, however, the code behaves simmilar to the single species version.  The values of the mass and energy do change a little but this is to be expected as there is no conservation.