Fixed coverage status report and Travis issue of reporting fails as passes
The coverage value reported is now only for the source files (the wrong directory was being reported in Makefile.am) and Travis successfully repoted a failed test as an actual Fail in the previous commit.  The reason for that fail was that the input files and "true" moment data files were not tracked with git so were missing...  Everything should now pass correctly!