Updated FieldCalculations.cpp to handle mesh refinement
The potential and field are now calculating correctly when MeshRefinement is true.  The field integral functions were also updated but this may need to be double checked when advancing in time.  There was also a silly bug when setting the initial condition due to the code assuming a fixed value of Nx_loc, and this has also been fixed.