Fixed a couple of bugs
The previous edits were almost there, but the function to set the variable value of epsilon was never run.  It then turned out there were a couple of silly bugs but these have been fixed and the code seems to run as expected when no mesh refinement.  Still need to fix the bounds for changing epsilon value when MeshRefinement is True!