Code now calculates momentum and kinetic energy in space
Instead of calculating the bulk velocity and temperature in space and saving those values to files, the code now computes the momentum and kinetic energy and saves those values to files.  In other words, the values are multiplied by rho now.  NOTE: the results seem correct as an initial run but the formulae used for the values should be checked (in particular, velocity was dividing by 3 and at first it seems that part of the kinetic energy formula still needs to be multiplied by rho).