Doping Profile can now vary on different sides of the channel
Previously, the doping profile took the values of NL in the middle and NH on the left and right of the channel.  Now, however, the values are N_left on the left, N_center in the middle and N_right on the right.  After doing an initial test run for nt=100 time-steps with N_left = NH, N_center = NL and N_right = NH, the code seems to be behaving as expected.  Note that old input files with NH and NL instead of N_left, N_center and N_right will still work because the code will set N_left = NH, N_center = NL and N_right = NH in this case.