Fixed annoying GRVY messages and infinite loop in t
The GRVY messages when N_left, N_center & N_right cannot be found have been suppressed.  Also, in an effort to print the correct marginals, the t++ command was previously shifted but to a point where only the processor with rank 0 was updating it.  This has now  been fixed too, so that the program can succesfully exit.  Finally, it has been noticed that for some reason certain quantities are not being printed for all relevant time-steps (e.g. the macroscopic quantities in x-space).  These values are now printed through use of a tmp variable, but this may not be enough to fix that problem...